services/capsula/requirements.txt
torch
torchvision
PyYAML==6.0.1
requests==2.31.0
tqdm==4.66.1
Pillow==10.2.0
Pygments==2.18.0
scipy>=1.4.1
pandas==2.2.0
seaborn==0.13.2
scikit-learn==1.4.0
scikit-multilearn==0.2.0
torchmetrics==1.8.2
torchmetrics[detection]
ultralytics==8.2.49
timm==0.9.12
onnx==1.17.0
boto3==1.37.20
safetensors>=0.4.0
tokenizers>=0.13.0
huggingface-hub==0.35.3
accelerate>=0.20.0
peft==0.17.1
transformers==4.56.2
albumentations==2.0.8
numpy<2.0.0
evaluate==0.4.1
pydantic==2.9.2
python-multipart==0.0.9
uvicorn==0.27.1
fastapi==0.109.2
mlflow==2.21.1
mkdocs==1.6.0
mkdocstrings==0.25.1
mkdocstrings-python==1.10.3
mkdocs-material==9.5.26
pytest==8.4.1
